feat： 添加 data_base 模块，用于处理音乐搜索和图像生成

+++ src/plugins/cloud_music.py
@@ -1,7 +1,14 @@
 import asyncio
 import pickle
-from pathlib import Path
+import os
 import time
+import uuid
+import requests
+from pathlib import Path
+from datetime import datetime
+from typing import Tuple, Optional
+from src.clover_music.cloud_music.data_base import netease_music_search_info_img,netease_music_info_img
+from src.configs.path_config import temp_path
 from nonebot import on_command
 from nonebot.rule import to_me
 from nonebot.exception import FinishedException
@@ -15,16 +22,17 @@ __name__ = "plugins | cloud_music"
 unikey_cache = {'unikey': None, 'expires': 0}
 
 music = on_command("点歌", rule=to_me(), priority=10, block=False)
-async def login(session):
-    """处理登录逻辑"""
-    if not os.path.exists('cloud_music_cookies.cookie'):
-        with open('cloud_music_cookies.cookie', 'wb') as f:
-            pickle.dump(session.cookies, f)
-    session.cookies = pickle.load(open('cloud_music_cookies.cookie', 'rb'))
-    session, status, user_id = await netease_cloud_music_is_login(session)
-    return session, status, user_id
 
-async def handle_qr_login(session):
+async def login(session: requests.Session) -> Tuple[requests.Session, bool, Optional[str]]:
+    """处理登录逻辑"""
+    cookie_file = 'cloud_music_cookies.cookie'
+    if not os.path.exists(cookie_file):
+        with open(cookie_file, 'wb') as f:
+            pickle.dump(session.cookies, f)
+    session.cookies = pickle.load(open(cookie_file, 'rb'))
+    return await netease_cloud_music_is_login(session)
+
+async def handle_qr_login(session: requests.Session) -> None:
     """处理二维码登录逻辑"""
     if unikey_cache['unikey'] and time.time() < unikey_cache['expires']:
         unikey = unikey_cache['unikey']
@@ -49,37 +57,55 @@ async def handle_qr_login(session):
         pickle.dump(session.cookies, f)
 
 @music.handle()
-async def handle_function(msg: MessageEvent):
+async def handle_function(msg: MessageEvent) -> None:
     try:
-        keyword = msg.get_plaintext().replace("/点歌", "").strip(" ")
-        if keyword == "":
-            await music.finish("\n请输入“/点歌+歌曲名”喔🎶")
-
+        values = msg.get_plaintext().removeprefix("/点歌").strip().split()
         session = requests.session()
         session, status, user_id = await login(session)
+        
         if not status:
             await music.send("登录失效，请联系管理员进行登录")
             await handle_qr_login(session)
+            return
 
-        song_id, song_name, singer, song_url = await netease_music_search(keyword, session)
-        song_name = str(song_name).replace(".", "·").replace("/", "、")
-        if song_id is None:
-            await music.finish("\n没有找到歌曲，或检索到的歌曲均为付费喔qwq\n这绝对不是我的错，绝对不是！")
+        if not values or not all(values):
+            await music.finish("\n请输入“/点歌+歌曲名”喔🎶")
+            return
 
-        await music.send(MessageSegment.text(f" 来源：网易云音乐\n歌曲：{song_name} - {singer}\n请稍等喔~正在下载中..."))
-        output_silk_path = await netease_music_download(song_id, song_name, singer, session)
+        keyword = values[0]
+        temp_file = os.path.join(temp_path, f"{datetime.now().date()}_{keyword}_{uuid.uuid4().hex}.png")
 
-        if output_silk_path == -1:
-            await music.send("歌曲音频获取失败：登录信息失效。")
-        elif output_silk_path is None:
-            await music.send("歌曲音频获取失败了Σヽ(ﾟД ﾟ; )ﾉ，请重试。")
-        else:
-            await music.send(MessageSegment.file_audio(Path(output_silk_path)))
-        await delete_file(output_silk_path)
+        if len(values) == 1:
+            await netease_music_search_info_img(keyword, session, temp_file)
+            await music.send(MessageSegment.file_image(Path(temp_file)))
+
+        if len(values) == 2:
+            idx = values[1]
+            if not idx.isdigit() or int(idx) < 1 or int(idx) > 10:
+                await music.finish("\n序号必须是数字且在1-10范围内喔qwq")
+            song_id, song_name, singer = await netease_music_info_img(keyword, session, idx, temp_file)
+            
+            if song_id is None:
+                await music.finish("\n没有找到歌曲，或检索到的歌曲均为付费喔qwq\n这绝对不是我的错，绝对不是！")
+            await music.send(MessageSegment.file_image(Path(temp_file)))
+            output_silk_path = await netease_music_download(song_id, song_name, singer, session)
+            
+            if output_silk_path == -1:
+                await music.send("歌曲音频获取失败：登录信息失效。")
+            elif output_silk_path is None:
+                await music.send("歌曲音频获取失败了Σヽ(ﾟД ﾟ; )ﾉ，请重试。")
+            else:
+                await music.send(MessageSegment.file_audio(Path(output_silk_path)))
+                await delete_file(output_silk_path)
+
+        await delete_file(temp_file)
         await music.finish()
 
+    except FinishedException:
+        raise
     except Exception as e:
-        if isinstance(e, FinishedException):
-            raise
         logger.error(f"处理点歌请求时发生错误: {e}")
-        await music.finish("处理点歌请求时发生错误，请稍后重试。这绝对不是我的错，绝对不是！")
+        r_msg = e.message
+        if r_msg == None:
+            r_msg = "未知错误，请稍后再试。"
+        await music.finish(f"处理点歌请求时发生错误：{r_msg}。这绝对不是我的错，绝对不是！")
